requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.135.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.0
fastar==0.8.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
psycopg==3.3.3
psycopg2==2.9.11
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.54.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
tomli==2.4.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
